refactor mcp proxy decision and tool helpers
